Tests for NH-2722 moved into linq-specified directory
