--- src/NHibernate.Test/NHSpecificTest/NH2722/Entity.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace NHibernate.Test.NHSpecificTest.NH2722
-{
-	public class Entity
-	{
-		public virtual Guid Id { get; set; }
-		public virtual int? Property { get; set; }
-	}
-}
